third_party/WebKit/LayoutTests/fast/mediastream/MediaStream-onactive-oninactive-expected.txt

Test onactive/oninactive callback.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS stream.active is true
PASS stream.getAudioTracks().length is 1
PASS stream.getVideoTracks().length is 1
PASS streamInactive was called.
PASS stream.active is false
PASS stream.active is false
PASS streamActive was called.
PASS stream.active is true
PASS successfullyParsed is true

TEST COMPLETE
